+++ app/controllers/comparisons/base_controller.rb
@@ -5,6 +5,8 @@ module Comparisons
     include UserTypeSpecific
     include ComparisonTableGenerator
     include SchoolGroupBreadcrumbs
+    include SchoolGroupAdvice
+    include SchoolGroupAccessControl
 
     skip_before_action :authenticate_user!
 
@@ -15,27 +17,13 @@ module Comparisons
     before_action :set_results, only: [:index]
     before_action :set_unlisted_schools_count, only: [:index]
     before_action :set_headers, only: [:index]
+    before_action :set_school_group_vars, only: [:index]
 
     protect_from_forgery except: :unlisted
 
     layout :switch_layout
 
     def index
-      # FIXME extract this so it can be shared with advice base controller
-      @school_group_layout = params[:group] == 'true'
-      if @school_group_layout
-        @school_group = SchoolGroup.find(params[:school_group_ids].reject(&:blank?).first)
-        @fuel_types = @school_group.fuel_types
-        @schools = @school_group.schools.active.accessible_by(current_ability, :show).by_name
-        @priority_action_count = SchoolGroups::PriorityActions.new(@schools).priority_action_count
-        @alert_count = SchoolGroups::Alerts.new(@schools).summarise.count
-        build_breadcrumbs([
-                            { name: I18n.t('advice_pages.breadcrumbs.root'), href: school_group_advice_path(@school_group) },
-                            { name: I18n.t('school_groups.titles.comparisons'), href: comparison_reports_school_group_advice_path(@school_group) },
-                            { name: @report.title }
-                          ])
-      end
-
       respond_to do |format|
         format.html do
           @charts = create_charts(@results)
@@ -177,7 +165,20 @@ module Comparisons
     end
 
     def switch_layout
-      params[:group_dashboard] == 'true' ? 'dashboards' : 'application'
+      params[:group] == 'true' ? 'dashboards' : 'application'
+    end
+
+    def set_school_group_vars
+      @school_group_layout = params[:group] == 'true'
+      return unless @school_group_layout
+      @school_group = SchoolGroup.find(params[:school_group_ids].reject(&:blank?).first)
+      set_all_group_advice_vars
+      redirect_unless_authorised and return
+      build_breadcrumbs([
+                          { name: I18n.t('advice_pages.breadcrumbs.root'), href: school_group_advice_path(@school_group) },
+                          { name: I18n.t('school_groups.titles.comparisons'), href: comparison_reports_school_group_advice_path(@school_group) },
+                          { name: @report.title }
+                        ])
     end
   end
 end
