+++ app/controllers/school_groups/advice/base_controller.rb
@@ -3,6 +3,7 @@ module SchoolGroups
     class BaseController < ApplicationController
       include SchoolGroupAccessControl
       include SchoolGroupBreadcrumbs
+      include SchoolGroupAdvice
 
       load_resource :school_group
 
@@ -27,20 +28,6 @@ module SchoolGroups
 
       def set_titles
         @page_title = t('page_title', scope: "school_groups.advice_pages.#{advice_page_key}", default: nil)
-      end
-
-      def set_fuel_types
-        @fuel_types = @school_group.fuel_types
-      end
-
-      # Rely on CanCan to filter the list of schools to those that can be shown to the current user
-      def load_schools
-        @schools = @school_group.schools.active.accessible_by(current_ability, :show).by_name
-      end
-
-      def set_counts
-        @priority_action_count = SchoolGroups::PriorityActions.new(@schools).priority_action_count
-        @alert_count = SchoolGroups::Alerts.new(@schools).summarise.count
       end
 
       def breadcrumbs
